+++ src/types/bee.ts
@@ -1221,7 +1221,30 @@ export type BeePluginAdvancedPermission = RecursivePartial<{
         hideOnMobile: AdvancedSettingsShowLocked
         hideOnAmp: AdvancedSettingsShowLocked
         id: AdvancedSettingsShowLocked
-
+      }
+    },
+    table: {
+      behaviors: AdvancedSettingsBehaviours
+      properties: {
+        columns: AdvancedSettingsShowLocked
+        rows: AdvancedSettingsShowLocked
+        backgroundColor: AdvancedSettingsShowLocked
+        border: AdvancedSettingsShowLocked
+        textColor: AdvancedSettingsShowLocked
+        linkColor: AdvancedSettingsShowLocked
+        fontFamily: AdvancedSettingsShowLocked
+        fontWeight: AdvancedSettingsShowLocked
+        fontSize: AdvancedSettingsShowLocked
+        textAlign: AdvancedSettingsShowLocked
+        lineHeight: AdvancedSettingsShowLocked
+        letterSpacing: AdvancedSettingsShowLocked
+        direction: AdvancedSettingsShowLocked
+        padding: AdvancedSettingsShowLocked
+        hideOnMobile: AdvancedSettingsShowLocked
+        hideOnAmp: AdvancedSettingsShowLocked
+        id: AdvancedSettingsShowLocked
+        alternateRowBackgroundColor: AdvancedSettingsShowLocked
+        headers: AdvancedSettingsShowLocked
       }
     }
     addon: {